Started moving non-dcat stuff out of dcat.
