+++ magda-csv-connector/aspect-templates/provenance.js
@@ -1,14 +1,5 @@
 const { fuzzy } = libraries;
 
-let isInternallyProduced = fuzzy.findClosestFieldThreshold(
-    dataset,
-    0.5,
-    "created internally"
-);
-if (isInternallyProduced) {
-    isInternallyProduced = fuzzy.similarity(isInternallyProduced, "yes") > 0.5;
-}
-
 let mechanism = fuzzy.findClosestFieldThreshold(
     dataset,
     0.5,
@@ -47,7 +38,6 @@ if (affiliatedOrganisation) {
 let provenance = undefined;
 
 if (
-    isInternallyProduced !== undefined ||
     mechanism !== undefined ||
     sourceSystem !== undefined ||
     likelihoodOfRelease !== undefined ||
@@ -55,12 +45,11 @@ if (
     affiliatedOrganisation
 ) {
     provenance = {
-        isInternallyProduced,
         mechanism,
         sourceSystem,
         likelihoodOfRelease,
         isOpenData,
-        affiliatedOrganisation
+        affiliatedOrganisationId: affiliatedOrganisation
     };
 }
