+++ magda-csv-connector/src/test/csv1.json
@@ -52,12 +52,6 @@
                 "accessLevel": "Non-Public",
                 "importance": "Essential",
                 "status": "Current",
-                "informationSecurity": {
-                    "disseminationLimits": [
-                        "Sensitive: Personal",
-                        "Sensitive: Commercial"
-                    ]
-                },
                 "metadata": {
                     "format": "META FMT"
                 }
@@ -79,6 +73,12 @@
                 "sourceSystem": "CRM",
                 "likelihoodOfRelease": "Low",
                 "affiliatedOrganisation": ["AI"]
+            },
+            "information-security": {
+                "disseminationLimits": [
+                    "Sensitive: Personal",
+                    "Sensitive: Commercial"
+                ]
             }
         },
         "sourceTag": "stag"
@@ -136,10 +136,7 @@
                 "contactPoint": "Engineering",
 
                 "accessLevel": "Restricted Public",
-                "importance": "Useful",
-                "informationSecurity": {
-                    "disseminationLimits": ["For Official Use Only"]
-                }
+                "importance": "Useful"
             },
             "source": {
                 "type": "csv-dataset",
@@ -156,6 +153,9 @@
             "provenance": {
                 "mechanism": "Data generated and provided to us by Data61",
                 "likelihoodOfRelease": "Restricted"
+            },
+            "information-security": {
+                "disseminationLimits": ["For Official Use Only"]
             }
         },
         "sourceTag": "stag"
@@ -211,9 +211,6 @@
                 "contactPoint": "META STD\n\nGaming",
                 "accessLevel": "T3",
                 "status": "Retired",
-                "informationSecurity": {
-                    "disseminationLimits": ["For Official Use Only"]
-                },
                 "metadata": {
                     "standard": "META STD"
                 }
@@ -235,6 +232,9 @@
                 "sourceSystem": "Recycle bin",
                 "likelihoodOfRelease": "Low",
                 "affiliatedOrganisation": ["Machine Learning"]
+            },
+            "information-security": {
+                "disseminationLimits": ["For Official Use Only"]
             }
         },
         "sourceTag": "stag"
