+++ magda-csv-connector/src/test/csv2.json
@@ -116,11 +116,7 @@
                 "keywords": ["KE1", "K W 2", "KW3", "Item", "Resource"],
                 "contactPoint": "Fred Nerk\n\nActive",
                 "accessLevel": "Not Classified",
-                "status": "Active",
-                "informationSecurity": {
-                    "disseminationLimits": ["Sensitive", "Sensitive: Personal"],
-                    "classification": "UNCLASSIFIED"
-                }
+                "status": "Active"
             },
             "source": {
                 "type": "csv-dataset",
@@ -140,10 +136,13 @@
                 "publisher": "org-connector-a94089507837f9062056995b14324baa"
             },
             "provenance": {
-                "isInternallyProduced": false,
                 "mechanism": "Sourced from a chicken farm",
                 "sourceSystem": "Sourced from a chicken farm",
                 "isOpenData": false
+            },
+            "information-security": {
+                "disseminationLimits": ["Sensitive", "Sensitive: Personal"],
+                "classification": "UNCLASSIFIED"
             }
         },
         "sourceTag": "stag"
@@ -221,10 +220,7 @@
                 "keywords": ["KW4", "Item", "Location"],
                 "contactPoint": "Joe Bloggs",
                 "accessLevel": "Not Classified",
-                "status": "Active",
-                "informationSecurity": {
-                    "classification": "UNCLASSIFIED"
-                }
+                "status": "Active"
             },
             "source": {
                 "type": "csv-dataset",
@@ -242,10 +238,12 @@
                 "publisher": "org-connector-634437f607a8f344c444482c3c831cc5"
             },
             "provenance": {
-                "isInternallyProduced": true,
                 "mechanism": "Internally produced by Agency",
                 "sourceSystem": "Internally produced by Agency",
                 "isOpenData": true
+            },
+            "information-security": {
+                "classification": "UNCLASSIFIED"
             }
         },
         "sourceTag": "stag"
@@ -323,11 +321,7 @@
                 "keywords": ["Item", "Resource"],
                 "contactPoint": "Active",
                 "accessLevel": "Not Classified",
-                "status": "Active",
-                "informationSecurity": {
-                    "disseminationLimits": ["Sensitive: Personal"],
-                    "classification": "PROTECTED"
-                }
+                "status": "Active"
             },
             "source": {
                 "type": "csv-dataset",
@@ -345,10 +339,13 @@
                 "publisher": "org-connector-e44009afeecf1fe914c6c84488c4280a"
             },
             "provenance": {
-                "isInternallyProduced": true,
                 "mechanism": "Produced with magic",
                 "sourceSystem": "Produced with magic",
                 "isOpenData": true
+            },
+            "information-security": {
+                "disseminationLimits": ["Sensitive: Personal"],
+                "classification": "PROTECTED"
             }
         },
         "sourceTag": "stag"
