+++ magda-registry-aspects/information-security.schema.json
@@ -1,0 +1,33 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Metadata about information security",
+    "type": "object",
+    "properties": {
+        "disseminationLimits": {
+            "title": "Any sensitivity DLMs. See https://www.protectivesecurity.gov.au/informationsecurity/Documents/AustralianGovernmentclassificationsystem.pdf for more details.",
+            "type": "array",
+            "items": {
+                "type": "string",
+                "enum": [
+                    "For Official Use Only",
+                    "Sensitive",
+                    "Sensitive: Personal",
+                    "Sensitive: Legal",
+                    "Sensitive: Cabinet",
+                    "Sensitive: Commercial"
+                ]
+            }
+        },
+        "classification": {
+            "title": "Security classification. See https://www.protectivesecurity.gov.au/informationsecurity/Documents/AustralianGovernmentclassificationsystem.pdf for more details.",
+            "type": "string",
+            "enum": [
+                "UNCLASSIFIED",
+                "PROTECTED",
+                "CONFIDENTIAL",
+                "SECRET",
+                "TOP SECRET"
+            ]
+        }
+    }
+}
