+++ magda-scala-common/src/main/scala/au/csiro/data61/magda/model/Registry.scala
@@ -196,6 +196,7 @@ trait RegistryConverters extends RegistryProtocols {
         )
       case _ => None
     }
+    val provenanceOpt = hit.aspects.get("provenance")
 
     val qualityAspectOpt = hit.aspects.get("dataset-quality-rating")
 
@@ -331,14 +332,8 @@ trait RegistryConverters extends RegistryProtocols {
       hasQuality = hasQuality,
       score = None,
       source = hit.aspects.get("source").map(_.convertTo[DataSouce]),
-      creation = dcatStrings
-        .getFields("creation")
-        .headOption
-        .filter {
-          case JsNull => false
-          case _      => true
-        }
-        .map(_.convertTo[DcatCreation]),
+      creation = provenanceOpt
+        .map(_.convertTo[Provenance]),
       publishingState = Some(
         publishing.extract[String]('state.?).getOrElse("published")
       ), // assume not set means published
@@ -568,7 +563,7 @@ object Registry extends RegistryConverters {
 
     def isRecordAspectEvent =
       this == EventType.CreateRecordAspect || this == EventType.DeleteRecordAspect || this == EventType.PatchRecordAspect
-        
+
     def isCreateEvent =
       this == EventType.CreateRecord || this == EventType.CreateRecordAspect || this == EventType.CreateAspectDefinition
